fix: resolve clang-tidy CI failures in ct_core
- Remove readability-identifier-naming (MemberPrefix m_ conflicted with
  public struct members; camelCase values were invalid config)
- Add std::uint8_t base type to FilterType enum (performance-enum-size)
- Cast std::size_t k/j/n to float in DFT angle expressions
  (cppcoreguidelines-narrowing-conversions)

+++ ct_core/src/FilterBank.cpp
@@ -12,7 +12,7 @@ static std::vector<std::complex<float>> dft(std::span<const float> x) {
     std::vector<std::complex<float>> X(n);
     for (std::size_t k = 0; k < n; ++k) {
         for (std::size_t j = 0; j < n; ++j) {
-            float angle = -2.f * static_cast<float>(std::numbers::pi) * k * j / n;
+            float angle = -2.f * static_cast<float>(std::numbers::pi) * static_cast<float>(k) * static_cast<float>(j) / static_cast<float>(n);
             X[k] += x[j] * std::complex<float>{std::cos(angle), std::sin(angle)};
         }
     }
@@ -24,7 +24,7 @@ static void idft(std::vector<std::complex<float>>& X, std::span<float> out) {
     for (std::size_t j = 0; j < n; ++j) {
         std::complex<float> sum{};
         for (std::size_t k = 0; k < n; ++k) {
-            float angle = 2.f * static_cast<float>(std::numbers::pi) * k * j / n;
+            float angle = 2.f * static_cast<float>(std::numbers::pi) * static_cast<float>(k) * static_cast<float>(j) / static_cast<float>(n);
             sum += X[k] * std::complex<float>{std::cos(angle), std::sin(angle)};
         }
         out[j] = sum.real() / static_cast<float>(n);
